tests/wpt/tests/html/semantics/scripting-1/the-script-element/execution-timing/150-import-xhtml.xhtml

<?xml version="1.0" encoding="utf-8"?>
<html xmlns="http://www.w3.org/1999/xhtml">
<head>
  <title>Stylesheet in XHTML BODY with @import blocking scripts</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script src="testlib/testlib.js"></script>
</head>
<body>
  <div id="log">FAILED (This TC requires JavaScript enabled)</div>
  <div id="test">Test</div>
  <!-- this stylesheet blocks scripts -->
  <link rel="stylesheet" href="css/import.css?pipe=trickle(d2)" />
  <script>
    test(function() {
      assert_equals(getComputedStyle(document.getElementById("test")).position,
                    "fixed");
    });
  </script>
</body>
</html>
