LayoutTests/imported/w3c/web-platform-tests/html/semantics/scripting-1/the-script-element/execution-timing/150-import-xhtml-expected.txt

FAILED (This TC requires JavaScript enabled)

FAIL Stylesheet in XHTML BODY with @import blocking scripts assert_equals: expected "fixed" but got "static"
Test
